+++ resources/components/laraknife/layout/col-empty.blade.php
@@ -1,4 +1,4 @@
 @props(['width' => '1'])
 <div class="col-md-{{$width}}">
-    <span class="pseudo-label"></span>
+    <span class="pseudo-label">&nbsp;</span>
 </div>